// src/app/models/People.ts
export interface People {
    name: string;
    last_name: string;
    first_name: string;
    age: string;
    height: string;
    mass: string;
    hair_color: string;
    skin_color: string;
    eye_color: string;
    birth_year: string;
    gender: string;
    homeworld: string;
    home_planet: string;
    this_film: string;
    // species:string;
    // films:string;
    // vehicles:string;
    // starships:string;
    // created:string;
    // edited:string;
    // url:string;
}